requirements.txt
# Shipwreck Excavation QGIS Plugin Requirements
# Core dependencies for the plugin functionality
# Install with: pip install -r requirements.txt

# Database dependencies (NEW - for Supabase)
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Image processing
Pillow>=9.0.0

# PDF generation and QR codes
reportlab>=3.6.0
qrcode>=7.3.0

# Video processing (optional but recommended)
opencv-python>=4.5.0

# Excel file reading
pandas>=1.3.0
openpyxl>=3.0.0

# Telegram bot integration (for field data collection)
python-telegram-bot>=20.0
aiofiles>=23.0.0

# Cloud storage dependencies (optional)
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# 3D visualization (optional)
vtk>=9.0.0

# Database
# sqlite3 is included with Python

# Additional utilities
numpy>=1.21.0  # Required by opencv-python
python-dateutil>=2.8.0
requests>=2.28.0